+++ bootcamp-milestone-2/src/app/blog/page.tsx
@@ -1,27 +1,8 @@
 
 import blogStyles from "@/app/blog/blog.module.css"
 
-export interface Blog {
-    title: string,
-    date: string,
-    description: string,
-    slug: string
-}
+import {Blog, myList} from "./blogData";
 
-const myList: Blog[] = [
-    {
-        title: "First time using TypeScript!",
-        date: "October 18th, 2023",
-        description: "After a challenging process of setting up Node and TypeScript, we're up and running.",
-        slug: "@/app/blog"
-    },
-    {
-        title: "Sitting in my bed coding!",
-        date: "October 18th, 2023",
-        description: "At least my lovely GF is here with me.",
-        slug: "@/app/blog"
-    }
-]
 
 
 function BlogPost ({title, date, description, slug}: Blog) {
